+++ lib/blazer/adapters/spark_adapter.rb
@@ -1,0 +1,6 @@
+module Blazer
+  module Adapters
+    class SparkAdapter < HiveAdapter
+    end
+  end
+end